New classes CharPointer_UTF8, CharPointer_UTF16, CharPointer_UTF32, complete refactoring of CharacterFunctions class and updates to the internals of String methods. Removal of String::operator+= (unsigned int) because of clashes with wide-char types that use unsigned int. Made core classes compatible with Android. Minor fixes to ListBox and AudioDeviceManager.

+++ src/native/linux/juce_linux_Network.cpp
@@ -358,7 +358,7 @@ private:
 
         header << "\r\nUser-Agent: JUCE/" << JUCE_MAJOR_VERSION << '.' << JUCE_MINOR_VERSION
                << "\r\nConnection: Close\r\nContent-Length: "
-               << postData.getSize() << "\r\n"
+               << (int) postData.getSize() << "\r\n"
                << headers << "\r\n";
 
         MemoryBlock mb;
